Scaffolds for core models: home_page and general_page
