+++ foundation_cms/base/models.py
@@ -1,6 +1,6 @@
 from .models.abstract_article_page import AbstractArticlePage
 from .models.abstract_base_page import AbstractBasePage, Author
-from .models.abstract_collection_page.py import AbstractCollectionPage
+from .models.abstract_collection_page import AbstractCollectionPage
 from .models.abstract_general_page import AbstractGeneralPage
 from .models.abstract_home_page import AbstractHomePage
 from .models.base_block import BaseBlock
